+++ src/base/state.hpp
@@ -10,7 +10,6 @@
 
 #include "framework/json.hpp"
 #include "framework/operations.hpp"
-#include "framework/circuit.hpp"
 #include "framework/types.hpp"
 #include "framework/data.hpp"
 #include "framework/creg.hpp"
@@ -29,10 +28,10 @@ public:
   using ignore_argument = void;
   State() = default;
   virtual ~State() = default;
-  
+
   //=======================================================================
   // Subclass Override Methods
-  // 
+  //
   // The following methods should be implemented by any State subclasses.
   // Abstract methods are required, while some methods are optional for
   // State classes that support measurement to be compatible with a general
@@ -41,12 +40,15 @@ public:
 
   //-----------------------------------------------------------------------
   // Abstract methods
-  // 
+  //
   // The implementation of these methods must be defined in all subclasses
   //-----------------------------------------------------------------------
   
+  // Return a string name for the State type
+  virtual std::string name() const = 0;
+
   // Return the set of qobj instruction types supported by the State
-  // by the Operations::OpType enum class. 
+  // by the Operations::OpType enum class.
   // Standard OpTypes that can be included here are:
   // - `OpType::gate` if gates are supported
   // - `OpType::measure` if measure is supported
@@ -58,7 +60,7 @@ public:
   // For the case of gates the specific allowed gates are checked
   // with the `allowed_gates` function.
   virtual std::unordered_set<Operations::OpType> allowed_ops() const = 0;
-  
+
   // Return the set of qobj gate instruction names supported by the state class
   // For example this could include {"u1", "u2", "u3", "U", "cx", "CX"}
   virtual stringset_t allowed_gates() const = 0;
@@ -83,7 +85,7 @@ public:
   // Initializes the State to a specific state.
   virtual void initialize_qreg(uint_t num_qubits, const state_t &state) = 0;
 
-  // Return an estimate of the required memory for implementing the 
+  // Return an estimate of the required memory for implementing the
   // specified sequence of operations on a `num_qubit` sized State.
   virtual uint_t required_memory_mb(uint_t num_qubits,
                                     const std::vector<Operations::Op> &ops) = 0;
@@ -97,14 +99,14 @@ public:
 
   //-----------------------------------------------------------------------
   // Optional: measurement sampling
-  // 
+  //
   // This method is only required for a State subclass to be compatible with
   // the measurement sampling optimization of a general the QasmController
   //-----------------------------------------------------------------------
 
   // Sample n-measurement outcomes without applying the measure operation
   // to the system state. Even though this method is not marked as const
-  // at the end of sample the system should be left in the same state 
+  // at the end of sample the system should be left in the same state
   // as before sampling
   virtual std::vector<reg_t> sample_measure(const reg_t &qubits,
                                             uint_t shots,
@@ -118,17 +120,17 @@ public:
   //=======================================================================
 
   //-----------------------------------------------------------------------
-  // Circuit validation
+  // OpSet validation
   //-----------------------------------------------------------------------
 
-  // Return false if circuit contains an unsupported instruction for
+  // Return false if an OpSet contains unsupported instruction for
   // the state class. Otherwise return true.
-  bool virtual validate_circuit(const Circuit &circ) const;
+  virtual bool validate_opset(const Operations::OpSet& opset) const;
 
-  // Raise an exeption if the circuit contains unsupported
-  // instructions for the state class. The exception message 
+  // Raise an exeption if the OpSet contains unsupported
+  // instructions for the state class. The exception message
   // contains the name of the unsupported instructions.
-  void virtual validate_circuit_except(const Circuit &circ) const;
+  virtual std::string invalid_opset_message(const Operations::OpSet& opset) const;
 
   //=======================================================================
   // Standard non-virtual methods
@@ -221,24 +223,33 @@ std::vector<reg_t> State<state_t>::sample_measure(const reg_t &qubits,
 
 
 template <class state_t>
-bool State<state_t>::validate_circuit(const Circuit &circ) const {
-  return circ.check_ops(allowed_ops(),
+bool State<state_t>::validate_opset(const Operations::OpSet &opset) const {
+  return opset.validate(allowed_ops(),
                         allowed_gates(),
                         allowed_snapshots());
 }
 
 
+
 template <class state_t>
-void State<state_t>::validate_circuit_except(const Circuit &circ) const {
+std::string State<state_t>::invalid_opset_message(const Operations::OpSet &opset) const {
   // Check operations are allowed
-  const auto invalid = circ.invalid_ops(allowed_ops(),
-                                        allowed_gates(),
-                                        allowed_snapshots());
-  if (!invalid.empty()) {
-    std::stringstream ss;
-    ss << "Circuit contains invalid instructions: " << invalid;
-    throw std::invalid_argument(ss.str());
-  }
+  auto invalid_optypes = opset.invalid_optypes(allowed_ops());
+  auto invalid_gates = opset.invalid_gates(allowed_gates());
+  auto invalid_snapshots = opset.invalid_snapshots(allowed_snapshots());
+  bool bad_instr = !invalid_optypes.empty();
+  bool bad_gates = !invalid_gates.empty();
+  bool bad_snaps = !invalid_snapshots.empty();
+  std::stringstream ss;
+  if (bad_gates)
+    ss << " invalid gate instructions: " << invalid_gates;
+  if (bad_snaps)
+    ss << " invalid snapshot instructions: " << invalid_snapshots;
+  // We can't print OpTypes so we add a note if there are invalid
+  // instructions other than gates or snapshots
+  if (bad_instr && (!bad_gates && !bad_snaps))
+    ss << " invalid non gate or snapshot instructions";
+  return ss.str();
 }